+++ app/utils/converters.ts
@@ -3,7 +3,7 @@ export const nowInSeconds = (): number => Math.floor(Date.now() / 1000);
 export const stringToNumber = (string: string): number => {
   const number = parseInt(string, 10);
 
-  if (isNaN(number)) {
+  if (Number.isNaN(number)) {
     return 0;
   }